+++ src/main/java/be/ugent/rml/cli/Main.java
@@ -2,13 +2,13 @@ package be.ugent.rml.cli;
 
 import be.ugent.rml.Executor;
 import be.ugent.rml.Utils;
+import be.ugent.rml.access.DatabaseType;
 import be.ugent.rml.conformer.MappingConformer;
 import be.ugent.rml.functions.FunctionLoader;
 import be.ugent.rml.functions.lib.IDLabFunctions;
 import be.ugent.rml.metadata.MetadataGenerator;
 import be.ugent.rml.records.RecordsFactory;
 import be.ugent.rml.store.QuadStore;
-import be.ugent.rml.store.QuadStoreFactory;
 import be.ugent.rml.store.RDF4JStore;
 import be.ugent.rml.store.SimpleQuadStore;
 import be.ugent.rml.term.NamedNode;