+++ src/main/java/be/ugent/rml/functions/FunctionUtils.java
@@ -1,11 +1,11 @@
 package be.ugent.rml.functions;
 
 import be.ugent.rml.NAMESPACES;
+import be.ugent.rml.Utils;
 import be.ugent.rml.store.Quad;
+import be.ugent.rml.store.QuadStore;
 import be.ugent.rml.term.NamedNode;
 import be.ugent.rml.term.Term;
-import be.ugent.rml.Utils;
-import be.ugent.rml.store.QuadStore;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -79,12 +79,26 @@ public class FunctionUtils {
         String typeStr = type.getValue();
 
         switch (typeStr) {
+            // This is quite crude, based on https://www.w3.org/TR/xmlschema11-2/#built-in-datatypes
             case "http://www.w3.org/2001/XMLSchema#string":
                 return String.class;
             case "http://www.w3.org/2001/XMLSchema#integer":
+            case "http://www.w3.org/2001/XMLSchema#long":
+            case "http://www.w3.org/2001/XMLSchema#int":
+            case "http://www.w3.org/2001/XMLSchema#short":
+            case "http://www.w3.org/2001/XMLSchema#byte":
+            case "http://www.w3.org/2001/XMLSchema#nonNegativeInteger":
+            case "http://www.w3.org/2001/XMLSchema#positiveInteger":
+            case "http://www.w3.org/2001/XMLSchema#unsignedLong":
+            case "http://www.w3.org/2001/XMLSchema#unsignedInt":
+            case "http://www.w3.org/2001/XMLSchema#unsignedShort":
+            case "http://www.w3.org/2001/XMLSchema#unsignedByte":
+            case "http://www.w3.org/2001/XMLSchema#nonPositiveInteger":
+            case "http://www.w3.org/2001/XMLSchema#negativeInteger":
                 return Integer.class;
             case "http://www.w3.org/2001/XMLSchema#decimal":
             case "http://www.w3.org/2001/XMLSchema#double":
+            case "http://www.w3.org/2001/XMLSchema#float":
                 return Double.class;
             case "http://www.w3.org/1999/02/22-rdf-syntax-ns#List":
                 return List.class;