+++ src/main/java/be/ugent/rml/records/ReferenceFormulationRecordFactory.java
@@ -5,6 +5,7 @@ import be.ugent.rml.store.QuadStore;
 import be.ugent.rml.term.Term;
 
 import java.io.IOException;
+import java.sql.SQLException;
 import java.util.List;
 
 /**
@@ -20,5 +21,5 @@ public interface ReferenceFormulationRecordFactory {
      * @return a list of records.
      * @throws IOException
      */
-    List<Record> getRecords(Access access, Term logicalSource, QuadStore rmlStore) throws IOException;
+    List<Record> getRecords(Access access, Term logicalSource, QuadStore rmlStore) throws IOException, SQLException, ClassNotFoundException;
 }