+++ src/test/resources/test-cases/RMLTC0001a-OracleDB/resource.sql
@@ -1,0 +1,2 @@
+CREATE TABLE student (Name VARCHAR(50));
+INSERT INTO student values ('Venus');