+++ src/test/resources/test-cases/RMLTC0007h-PostgreSQL/resource.sql
@@ -2,7 +2,6 @@ DROP TABLE IF EXISTS student;
 
 CREATE TABLE student (
   "ID" INTEGER,
-  "FirstName" VARCHAR(50),
-  "LastName" VARCHAR(50)
+  "Name" VARCHAR(50)
 );
-INSERT INTO student values ('10', 'Venus', 'Williams');
+INSERT INTO student values ('10', 'Venus');